validate ILT links exist, defer websocket until after resolution, rename file to match export
